refactor(router): extract logging, subscriptions, and notifications
Phase 2 of #1256, continuing from the task group. `router.rs` goes 4806 to
4635 lines; `router/notify.rs` holds the 14 methods.

Everything the server says to a client without being asked: `log` and its
four level helpers, the resource-subscription bookkeeping, and the
`notifications/*/list_changed` and `notifications/resources/updated` senders.

Subscriptions moved with them rather than staying beside the resource
registry, because they exist to drive `notify_resource_updated`. A
subscription nobody notifies is bookkeeping for its own sake.

An `impl McpRouter` block in a child module, so neither the type nor its API
changed. Twelve of the fourteen are `pub` and stay exactly as reachable as
before. The two private ones, `subscribe` and `unsubscribe`, become
`pub(super)` for the same reason as the task group: a child module's private
items are not visible to its parent.

1030 lib tests before and after. Boundaries asserted against the doc comment
opening the block and the item following it, plus the method count and how
many of them were private.

Refs #1256
